+++ SimpleAuth.Server/Controllers/UserController.cs
@@ -8,10 +8,8 @@ using SimpleAuth.Core.Extensions;
 using SimpleAuth.Repositories;
 using SimpleAuth.Server.Extensions;
 using SimpleAuth.Server.Middlewares;
-using SimpleAuth.Server.Services;
 using SimpleAuth.Services;
 using SimpleAuth.Services.Entities;
-using SimpleAuth.Shared;
 using SimpleAuth.Shared.Enums;
 using SimpleAuth.Shared.Exceptions;
 using SimpleAuth.Shared.Models;
@@ -26,16 +24,13 @@ namespace SimpleAuth.Server.Controllers
     public class UserController : BaseController<IUserService, IUserRepository, User>
     {
         private readonly IEncryptionService _encryptionService;
-        private readonly IGoogleService _googleService;
         private readonly IUserValidationService _userValidationService;
         private readonly ILogger<UserController> _logger;
 
-        public UserController(IServiceProvider serviceProvider, IEncryptionService encryptionService,
-            IGoogleService googleService, IUserValidationService userValidationService) : base(
+        public UserController(IServiceProvider serviceProvider, IEncryptionService encryptionService, IUserValidationService userValidationService) : base(
             serviceProvider)
         {
             _encryptionService = encryptionService;
-            _googleService = googleService;
             _userValidationService = userValidationService;
             _logger = serviceProvider.ResolveLogger<UserController>();
         }
@@ -117,71 +112,6 @@ namespace SimpleAuth.Server.Controllers
             });
         }
 
-        [HttpPost("_google")]
-        public async Task<IActionResult> CheckGoogleToken([FromBody] LoginByGoogleRequest form)
-        {
-            if (!ModelState.IsValid)
-                return BadRequest();
-
-            return await ProcedureDefaultResponseIfError(async () =>
-            {
-                var user = Service.GetUser(form.Email, RequestAppHeaders.Corp);
-                var localUserInfo = user?.LocalUserInfos?.FirstOrDefault(x => x.Corp == RequestAppHeaders.Corp);
-                if (localUserInfo == null)
-                    throw new EntityNotExistsException($"{form.Email} at {RequestAppHeaders.Corp}");
-
-                if (localUserInfo.Locked)
-                    throw new AccessLockedEntityException($"{user.Id} at {localUserInfo.Corp}");
-
-                GoogleTokenResponseResult ggToken;
-
-                try
-                {
-                    ggToken = await _googleService.GetInfoAsync(form.GoogleToken);
-                }
-                catch (SimpleAuthException ex)
-                {
-                    _logger.LogWarning(ex, $"Google token verification issue");
-                    return StatusCodes.Status412PreconditionFailed.WithMessage(ex.Message);
-                }
-
-                if (!form.Email.EqualsIgnoreCase(ggToken.Email))
-                    throw new DataVerificationMismatchException(
-                        $"This token belong to {ggToken.Email} which is different than email provided in payload {form.Email}"
-                    );
-
-                if (!form.VerifyWithClientId.IsBlank())
-                    if (!form.VerifyWithClientId.EqualsIgnoreCase(ggToken.Aud))
-                        throw new DataVerificationMismatchException(
-                            $"This token is rejected, it's not belong to client id {form.VerifyWithClientId}"
-                        );
-
-                if (!form.VerifyWithGSuite.IsBlank())
-                    if (!form.VerifyWithGSuite.EqualsIgnoreCase(ggToken.Hd))
-                        throw new DataVerificationMismatchException(
-                            $"This token is rejected, it's not belong to GSuite domain {form.VerifyWithGSuite}"
-                        );
-
-                var expiryDate = new DateTime(1970, 1, 1, 1, 1, 1, DateTimeKind.Utc);
-                expiryDate = expiryDate.AddSeconds(ggToken.Exp);
-
-                if (DateTime.UtcNow > expiryDate)
-                {
-                    _logger.LogWarning($"User {form.Email} provided an expired google token");
-                    return StatusCodes.Status406NotAcceptable.WithMessage(
-                        "Token already expired"
-                    );
-                }
-
-                var model = await GetBaseResponseUserModelAsync(form.Email);
-                model.GoogleToken = ggToken;
-                model.ExpireAt(expiryDate);
-
-                _logger.LogInformation($"Access granted for {form.Email} via Google Token");
-                return ReturnResponseUserModel(model);
-            });
-        }
-
         [HttpGet("{userId}/roles")]
         public async Task<IActionResult> GetActiveRoles(string userId)
         {
@@ -192,7 +122,7 @@ namespace SimpleAuth.Server.Controllers
         public async Task<IActionResult> GetUser(string userId)
         {
             return await ProcedureDefaultResponseIfError(() =>
-                GetBaseResponseUserModelAsync(userId).ContinueWith(x => ReturnResponseUserModel(x.Result))
+                GetBaseResponseUserModelAsync(userId, Service).ContinueWith(x => ReturnResponseUserModel(x.Result))
             );
         }
 
@@ -281,38 +211,6 @@ namespace SimpleAuth.Server.Controllers
                     );
                 }
             );
-        }
-
-        private async Task<ResponseUserModel> GetBaseResponseUserModelAsync(string userId)
-        {
-            var user = Service.GetUser(userId, RequestAppHeaders.Corp);
-            if (user == default)
-                throw new EntityNotExistsException(userId);
-
-            var filterRoleEnv = GetHeader(Constants.Headers.FilterByEnv);
-            var filterRoleTenant = GetHeader(Constants.Headers.FilterByTenant);
-
-            var activeRoles = await Service.GetActiveRolesAsync(userId, RequestAppHeaders.Corp, RequestAppHeaders.App,
-                filterRoleEnv, filterRoleTenant);
-            return new ResponseUserModel
-            {
-                Id = userId,
-                Corp = RequestAppHeaders.Corp,
-                ActiveRoles = activeRoles.OrEmpty().Select(x => new RoleModel
-                {
-                    Role = x.RoleId,
-                    Permission = x.Permission.Serialize()
-                }).ToArray()
-            };
-        }
-
-        private IActionResult ReturnResponseUserModel(ResponseUserModel model)
-        {
-            return (
-                model.ActiveRoles.IsAny()
-                    ? StatusCodes.Status200OK
-                    : StatusCodes.Status204NoContent
-            ).WithJson(model);
         }
     }
 }
